requirements.txt
numpy==1.23.5
pandas==1.5.1
scikit-learn==1.1.3
tensorflow==2.11.0
keras==2.11.0
torch==2.1.1
torchvision==0.16.1
matplotlib==3.6.2
seaborn==0.12.1

# Graph Network Libraries
torch-geometric==2.0.4
dgl==1.1.2

# Knowledge Graph Libraries
rdflib==6.1.1
owlready2==0.35
kglab==0.6.6
pykeen==1.10.1

# General Machine Vision Libraries
opencv-python==4.8.1.78
scikit-image==0.22.0
scipy==1.9.3

# Image Processing Libraries
pillow==9.3.0
imageio>=2.27

# Feature Detection and Matching Libraries
opencv-contrib-python==4.8.1.78

# Camera and Video Processing Libraries
pyzbar==0.1.9

# 3D Vision and Reconstruction Libraries
open3d==0.17.0
pymeshlab==2022.2.post4